backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
